requirements.txt
exa-py>=1.0.0
anthropic>=0.40.0
python-telegram-bot>=21.0
python-docx>=1.1.0
schedule>=1.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
